+++ src/render/mod.rs
@@ -5,6 +5,7 @@ pub mod factory;
 pub mod font;
 pub mod membrane;
 pub mod mirx_convert;
+pub mod mirx_frames;
 pub mod painter;
 pub mod partial;
 pub mod path;